Fix a few bugs in SessionLogin and login.php
This commit prevents SessionLogin to try to initialise the session when it
has already been initialised, causing a warning to be displayed.
It also removes an obsolete direct reference to $_SESSION['sessionLogin'] in
login.php (which was causing a warning as well since this variable does no
longer exist) and replaces this with a call to
SessionLogin::getUsername().
Also fix a small style inconsistency in checkIfInitialised().