Fix bug making login impossible
Fix the issue causing the login state of the user not being remembered
after the end of the script, causing the whole login system not to work.
The session_start() calls have been moved out from the SessionLogin class
as they need to be called before any output is generated and thus
they need to be put in the first script that is executed (in our case,
pages), so it move session_start() to login.php, and session_start()
should be put at the very start of any php page that needs to use the
login system.
Lastly, the login state of the user is no more stored as an object as this
caused issues (session variables can't hold references). It has been
instead replaced by a string representing the username of the user when
they're logged in or by a null value when they're not. Besides the obivous
advantage that it actually allows the system to work, it is more logical
(no username is stored when the user is not logged in), more efficient
(less classes to load, less operations and intricate functions), and more
readable (less code, less classes).