Improve branch creation
Improve the URL format to create a new branch. $_GET['id'] does not need
to be set to an id that is not taken by any branch to create a new branch.
Instead, to create a new branch, $_GET['id'] needs to be not set, so when
the user access manage-branches.php instead of
manage-branches.php?id=someid, they will get the form to create a new
branch.
Furthermore, when $_GET['id'] is not corresponding to any existing branch,
an error message is displayed.
Lastly, branch creation and branch update pages are now more distinct. The
'delete' button is removed from the branch creation page, for example.