Refactor model classes
Remove SQL commands from model classes and move the rest of the model
classes' code (except some methods) in store classes. This is to make
things simpler and more efficient, as BranchManager, SalesAssistant,
Customer and CompanyManager are now simple store classe which goal is only
to represent their corresponding entity and validate input. Furthermore,
since the SQL is now done by whatever script that needs it, the script can
optimize the query to its need.