Add transactions and replace MAX hack by last insert id in register.php
This allows for a code slightly easier to understand, but also for faster
and, more importantly, more robust SQL commands. The registration process
is now done using transactions, preventing the page to add a user and then
fail to insert the corresponding person, or to add a user, the
corresponding person but fail to insert the corresponding customer.
Furthermore, the use of the MAX SQL function to get the ID of the inserted
person has been replaced by the use of lastInsertId. This should prevent
critical problems arising due to the facts that at the time of the MAX
request, another person was added eventually linking the customer to the
wrong person.